horse_classification/requirements.txt
numpy==2.0.2
opencv-python==4.11.0.86
scipy==1.15.2
scikit-learn==1.6.1
scikit-image==0.25.2
Pillow==11.2.1
tensorflow==2.18.0
torch==2.6.0+cu124
ultralytics
pyyaml
joblib                       
matplotlib                   
tqdm
